+++ eng/package.ps1
@@ -1,6 +1,6 @@
 [CmdletBinding()]
 param(
-    [string]$Version = '0.1.0-beta.21',
+    [string]$Version = '0.1.0-beta.22',
     [string]$InnoCompiler,
     [string]$SbomTool
 )